Fix compiler warnings
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@6077 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/validator/libsres/ns_print.c
@@ -823,7 +823,7 @@ nxtbitmaps:
         int             n, m;
         char           *p;
 
-        len = SPRINTF((tmp, "\\# %u (\t; %s", edata - rdata, comment));
+        len = SPRINTF((tmp, "\\# %u (\t; %s", (unsigned)(edata - rdata), comment));
         T(addstr(tmp, len, &buf, &buflen));
         while (rdata < edata) {
             p = tmp;
